revert fdr_utils changes
